+++ common/db/mongo_notifications.go
@@ -2,11 +2,14 @@ package db
 
 import (
 	"context"
+	"encoding/json"
 	"time"
 
 	"go.mongodb.org/mongo-driver/v2/bson"
 	"go.mongodb.org/mongo-driver/v2/mongo"
 	"go.mongodb.org/mongo-driver/v2/mongo/options"
+
+	"asika/common/models"
 )
 
 func (s *mongoStorage) PutNotificationPrefs(username string, data []byte) error {
@@ -60,4 +63,32 @@ func (s *mongoStorage) DeleteNotificationDedup(key string) error {
 	defer cancel()
 	_, err := s.coll(BucketNotificationDedup).DeleteOne(ctx, bson.M{"_id": key})
 	return err
+}
+
+func (s *mongoStorage) ListNotificationPrefs(usernames []string) ([]models.NotificationPreferences, error) {
+	ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
+	defer cancel()
+	filter := bson.M{}
+	if len(usernames) > 0 {
+		filter["username"] = bson.M{"$in": usernames}
+	}
+	cursor, err := s.coll(BucketNotificationPrefs).Find(ctx, filter)
+	if err != nil {
+		return nil, err
+	}
+	defer cursor.Close(ctx)
+	var prefs []models.NotificationPreferences
+	for cursor.Next(ctx) {
+		var doc bson.M
+		if err := cursor.Decode(&doc); err != nil {
+			continue
+		}
+		data, _ := doc["data"].(string)
+		var p models.NotificationPreferences
+		if err := json.Unmarshal([]byte(data), &p); err != nil {
+			continue
+		}
+		prefs = append(prefs, p)
+	}
+	return prefs, nil
 }